+++ pkg/signals/signals.go
@@ -1,8 +1,8 @@
 package signals
 
 type NotePushed struct {
-	KeyCode  int `hubman:"key_code"`
-	Velocity int `hubman:"velocity"`
+	KeyCode  float64 `hubman:"key_code"`
+	Velocity float64 `hubman:"velocity"`
 }
 
 func (c NotePushed) Code() string {
@@ -14,8 +14,8 @@ func (c NotePushed) Description() string {
 }
 
 type NoteHold struct {
-	KeyCode  int `hubman:"key_code"`
-	Velocity int `hubman:"velocity"`
+	KeyCode  float64 `hubman:"key_code"`
+	Velocity float64 `hubman:"velocity"`
 }
 
 func (c NoteHold) Code() string {
@@ -27,8 +27,8 @@ func (c NoteHold) Description() string {
 }
 
 type NoteReleased struct {
-	KeyCode  int `hubman:"key_code"`
-	Velocity int `hubman:"velocity"`
+	KeyCode  float64 `hubman:"key_code"`
+	Velocity float64 `hubman:"velocity"`
 }
 
 func (c NoteReleased) Code() string {
@@ -42,8 +42,8 @@ func (c NoteReleased) Description() string {
 // ControlPushed В MIDI Control имеет только один тип событий "ControlChange",
 // поэтому длительность и конец нажатия здесь не отслеживаются
 type ControlPushed struct {
-	KeyCode int `hubman:"key_code"`
-	Value   int `hubman:"velocity"`
+	KeyCode float64 `hubman:"key_code"`
+	Value   float64 `hubman:"velocity"`
 }
 
 func (c ControlPushed) Code() string {
